requirements.txt
absl-py==2.3.0
accelerate==1.5.2
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
augly==1.0.0
av==15.0.0
basicsr==1.4.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.2
contourpy==1.3.2
cycler==0.12.1
decord==0.6.0
diffusers==0.35.0
easydict==1.13
einops==0.8.1
filelock==3.18.0
flash_attn @ file:///public/wangguanjie/miniconda3/envs/I2V/flash_attn-2.7.4.post1%2Bcu12torch2.6cxx11abiFALSE-cp311-cp311-linux_x86_64.whl#sha256=58853b28a5a926cae14402bfd8d4d93a45ebf8f9e79533f37ab09d0d77a99c05
fonttools==4.58.0
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
galois==0.4.6
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.73.1
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
lightning==2.5.1
lightning-utilities==0.14.2
llvmlite==0.44.0
lmdb==1.6.2
lpips==0.1.4
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.2.0
networkx==3.4.2
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.11.0.86
packaging==24.2
pandas==2.3.2
peft==0.17.0
pillow==11.1.0
piq==0.8.0
platformdirs==4.3.8
portalocker==3.2.0
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
ptvsd==4.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
qwen-vl-utils==0.0.11
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.34.1
six==1.17.0
smmap==5.0.2
sympy==1.13.1
tb-nightly==2.20.0a20250701
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.3.13
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchinfo==1.8.0
torchmetrics==1.7.0
torchsummary==1.5.1
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
triton==3.2.0
typing-inspection==0.4.1
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
wandb==0.21.0
wcwidth==0.2.13
Werkzeug==3.1.3
yapf==0.43.0
yarl==1.18.3
zipp==3.21.0
